+++ tests/testthat/test_write2.R
@@ -40,11 +40,11 @@ test_that("write2.freqlist -> HTML", {
 
 test_that("write2.freqlist -> doc", {
   expect_write2_worked(write2word, freqlist(table(mockstudy[, c("arm", "sex", "mdquality.s")], useNA = "ifany"), groupBy = c("arm", "sex")),
-                       reference = "write2.freqlist.doc.md", single = TRUE, caption = "My cool caption", format = "pandoc")
+                       reference = "write2.freqlist.doc.md", single = TRUE, title = "My cool title")
 })
 
 ###########################################################################################################
-#### External output
+#### External output, commented out on 11/9/17 because of external package changes
 ###########################################################################################################
 #
 # test_that("write2.knitr_kable -> HTML", {
@@ -154,7 +154,7 @@ test_that("write2.yaml -> PDF", {
 #             "tests/testthat/write2.freqlist.html", single = TRUE, render. = FALSE)
 #
 # write2word(freqlist(table(mockstudy[, c("arm", "sex", "mdquality.s")], useNA = "ifany"), groupBy = c("arm", "sex")),
-#            "tests/testthat/write2.freqlist.doc", single = TRUE, caption = "My cool caption", format = "pandoc", render. = FALSE)
+#            "tests/testthat/write2.freqlist.doc", single = TRUE, title = "My cool title", render. = FALSE)
 #
 #  write2html(knitr::kable(head(mockstudy)),
 #             "tests/testthat/write2.kable.html", render. = FALSE)